Flask/requirements.txt
alembic==1.16.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Faker==37.3.0
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
paho-mqtt==2.1.0
PyMySQL==1.1.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
simple-websocket==1.1.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.22.0
